feature: #71 default traverse specified dir
