+++ internal/fs/core.go
@@ -1,0 +1,7 @@
+package fs
+
+import "os"
+
+type FileTask struct {
+	info os.FileInfo
+}
