+++ internal/utils/utils.go
@@ -5,32 +5,12 @@ import (
 	"fmt"
 	"log"
 	"os"
-	"path/filepath"
-	"runtime"
 	"strconv"
 	"strings"
 	"time"
 
 	"github.com/fatih/color"
 )
-
-type FileTask struct {
-	info os.FileInfo
-}
-
-// Helper function to expand tilde in path
-func ExpandTilde(path string) string {
-	if !strings.HasPrefix(path, "~") {
-		return path
-	}
-
-	home, err := os.UserHomeDir()
-	if err != nil {
-		return path
-	}
-
-	return filepath.Join(home, path[1:])
-}
 
 func FormatSize(bytes int64) string {
 	const (
@@ -166,101 +146,4 @@ func LogDeletionToFile(files map[string]string) {
 		return
 	}
 	defer file.Close()
-}
-
-func DeleteEmptySubfolders(dir string) {
-	emptyDirs := make([]string, 0)
-
-	filepath.WalkDir(dir, func(path string, info os.DirEntry, err error) error {
-		if info == nil && !info.IsDir() {
-			return nil
-		}
-
-		if isEmptyDir(path) {
-			emptyDirs = append(emptyDirs, path)
-		}
-
-		return nil
-	})
-
-	for i := len(emptyDirs) - 1; i >= 0; i-- {
-		os.Remove(emptyDirs[i])
-	}
-}
-
-// Check if a directory is empty,true if directory have subfolders
-func isEmptyDir(dirPath string) bool {
-	dir, err := os.Open(dirPath)
-	if err != nil {
-		return false
-	}
-	defer dir.Close()
-
-	entries, err := dir.Readdir(0)
-
-	if err != nil {
-		return false
-	}
-	if len(entries) == 0 {
-		return true
-	}
-
-	for _, entry := range entries {
-		if entry.IsDir() {
-			// If this is a directory, we check recursively
-			if !isEmptyDir(filepath.Join(dirPath, entry.Name())) {
-				return false
-			}
-		} else {
-			return false
-		}
-	}
-	return true
-}
-
-func DeleteFiles(dir string, extensions []string, exclude []string, minSize int64) {
-	taskCh := make(chan FileTask, runtime.NumCPU())
-	filepath.Walk(dir, func(path string, info os.FileInfo, err error) error {
-		if info == nil {
-			return nil
-		}
-
-		if err != nil {
-			return nil
-		}
-
-		go func(path string, info os.FileInfo) {
-			// Acquire token from channel first
-			taskCh <- FileTask{info: info}
-			defer func() { <-taskCh }() // Release token when done
-
-			if len(exclude) != 0 {
-				for _, excludePattern := range exclude {
-					if strings.Contains(filepath.ToSlash(path), excludePattern+"/") ||
-						strings.HasPrefix(info.Name(), excludePattern) {
-						return
-					}
-				}
-			}
-
-			if len(extensions) > 0 {
-				ext := strings.ToLower(filepath.Ext(path))
-				matched := false
-				for _, allowedExt := range extensions {
-					if ext == allowedExt {
-						matched = true
-						break
-					}
-				}
-				if !matched {
-					return
-				}
-			}
-
-			if info.Size() > minSize {
-				os.Remove(path)
-			}
-		}(path, info)
-		return nil
-	})
 }
